requirements.txt
absl-py==2.1.0
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.1.0
blis==1.0.1
bokeh==3.6.0
catalogue==2.0.10
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.19.0
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dacite==1.8.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
dython==0.7.8
etelemetry==0.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fitz==0.0.1.dev2
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
holidays==0.58
htmlmin==0.1.12
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.0
hyperlink==21.0.0
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.4.5
incremental==24.7.2
inflect==7.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
isodate==0.6.1
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
lightgbm==4.5.0
lightning-utilities==0.11.9
llvmlite==0.43.0
loguru==0.7.2
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
ml_dtypes==0.5.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai==1.59.3
opencv-python==4.10.0.84
openpyxl==3.1.5
optree==0.13.0
opt_einsum==3.4.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandas-profiling==3.6.6
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
patsy==0.5.6
pefile==2023.2.7
phik==0.12.4
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pmdarima==2.0.4
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
prophet==1.1.6
Protego==0.3.1
protobuf==4.25.5
prov==2.0.1
psutil==6.0.0
puremagic==1.28
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
pydot==3.0.3
pygame==2.6.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pynndescent==0.5.13
pyOpenSSL==24.2.1
pyparsing==3.1.4
PyPDF2==3.0.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyqt5-tools==5.15.9.3.3
pyscroll==2.31
PySocks==1.7.1
pyspark==3.5.3
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyTMX==3.32
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyWavelets==1.7.0
pywin32==307
pywin32-ctypes==0.2.3
pywinpty==2.0.13
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
py-cpuinfo==9.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
queuelib==1.7.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
Scrapy==2.11.2
seaborn==0.13.2
selenium==4.25.0
Send2Trash==1.8.3
sentry-sdk==2.19.0
service-identity==24.1.0
setproctitle==1.3.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stanio==0.5.1
starlette==0.42.0
statsmodels==0.14.4
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.20.1
tokenize_rt==6.1.0
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.6.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.4.3
transformers==4.45.2
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.7.0
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.31
ultralytics-thop==2.0.11
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
visions==0.7.6
w3lib==2.2.1
wandb==0.18.7
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wikipedia==1.4.0
Wikipedia-API==0.7.1
win32-setctime==1.1.0
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.2
xmltodict==0.14.2
xyzservices==2024.9.0
yarl==1.18.3
ydata-profiling==4.11.0
yellowbrick==1.5
zope.interface==7.0.3
pip==23.0.1
setuptools==65.5.0
